refactor(PaymentCard): rename Types.ts to types.ts for consistency
All other types files in the codebase use lowercase types.ts.
This is a breaking change for users importing from
'@dnb/eufemia/extensions/payment-card/utils/Types' — update
the import path to lowercase 'types'.
