+++ packages/dnb-design-system-portal/src/docs/uilib/about-the-lib/releases/eufemia/v11-info.mdx
@@ -257,6 +257,7 @@ All import paths that changed or were removed in v11. Update these before changi
 | `style/themes/theme-carnegie/`                                              | `style/themes/carnegie/`                                                            | Path shortened                                |
 | `@dnb/eufemia/components/space/types` (`SectionSpacing`)                    | `@dnb/eufemia/components/space/types` (`InnerSpaceType`)                            | Type renamed                                  |
 | `export type { Props }` from any Field/Value/Form module                    | Use component-prefixed name (e.g. `FieldStringProps`)                               | See [Props Type Exports](#props-type-exports) |
+| `@dnb/eufemia/extensions/payment-card/utils/Types`                          | `@dnb/eufemia/extensions/payment-card/utils/types`                                  | Filename lowercased for consistency            |
 
 ### Common migration mistakes
