+++ packages/dnb-eufemia/src/extensions/payment-card/utils/types.ts
